Restructure Package
