+++ sharding-example/src/main/java/com/alpha/mongodb/sharding/example/configuration/CollectionShardedReactiveMongoConfiguration.java
@@ -1,7 +1,7 @@
 package com.alpha.mongodb.sharding.example.configuration;
 
+import com.alpha.mongodb.sharding.core.CollectionShardedReactiveMongoTemplate;
 import com.alpha.mongodb.sharding.core.configuration.CollectionShardingOptions;
-import com.alpha.mongodb.sharding.core.template.CollectionShardedReactiveMongoTemplate;
 import com.mongodb.ConnectionString;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
