+++ sharding-spring/src/main/java/com/alpha/mongodb/sharding/core/CompositeShardedReactiveMongoTemplate.java
@@ -1,4 +1,4 @@
-package com.alpha.mongodb.sharding.core.template;
+package com.alpha.mongodb.sharding.core;
 
 import com.alpha.mongodb.sharding.core.configuration.CompositeShardingOptions;
 import com.mongodb.reactivestreams.client.MongoClient;
