+++ sharding-spring/src/test/java/com/alpha/mongodb/sharding/core/CollectionShardedMongoTemplateTest.java
@@ -1,4 +1,4 @@
-package com.alpha.mongodb.sharding.core.template;
+package com.alpha.mongodb.sharding.core;
 
 import com.alpha.mongodb.sharding.core.callback.HintResolutionCallback;
 import com.alpha.mongodb.sharding.core.configuration.CollectionShardingOptions;
