+++ sharding-spring/src/test/java/com/alpha/mongodb/sharding/core/CompositeShardedMongoTemplateTest.java
@@ -1,4 +1,4 @@
-package com.alpha.mongodb.sharding.core.template;
+package com.alpha.mongodb.sharding.core;
 
 import com.alpha.mongodb.sharding.core.configuration.CompositeShardingOptions;
 import com.mongodb.client.MongoClient;
